Update changelog and function names
